requirements.txt
django==5.0.6
djangorestframework
environs